:recycle: Refactor: DocumentService interface화 - #15
